refactor(CompetitivitySection): 경쟁력 섹션의 카드 컴포넌트 분리
