+++ client/src/components/section/competitivity-section/competitivity-section.stories.tsx
