+++ client/src/components/section/competitivity-section/index.ts
