+++ client/src/components/section/index.ts
@@ -1,4 +1,4 @@
-export { CompetitivitySection } from './competitiveness-section';
+export { CompetitivitySection } from './competitivity-section';
 export { MenuSection } from './menu-section';
 export { FamilySection } from './family-section';
 export { FranchiseSection } from './franchise-section';
